refactor: Effect component を導入
